+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Added
+- Added an `OpenAIEmbedder` that uses the OpenAI Embeddings API to embed documents.
 
 
 ## [v1.1.1] - 2024-08-14